pyproject.toml
[project]
name = "data-orchestrator"
version = "0.1.0"
description = "Lightweight private data orchestration framework."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "apscheduler>=3.10",
    "tenacity>=8.0",
    "sqlalchemy>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "requests>=2.31",
]

[project.optional-dependencies]
ui = ["streamlit>=1.30"]
postgres = ["psycopg2-binary>=2.9"]
dev = ["pytest", "pytest-cov", "responses", "freezegun"]

[project.scripts]
orchestrator = "orchestrator.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
omit = ["orchestrator/ui/*"]
